Adds scratchpad mode; fixes a bug related to i3bar.
First, quickswitch now has a scratchpad mode, activated by the flag -s
or --scratchpad. This will cause it to list scratchpad windows instead
of regular ones, and subsequently `scratchpad show` the chosen window.

The bug happened when i3bar was not set to hide, and caused it to show
up in the output of the regular quickswitch as a window, which makes no
sense. It's now ignored, though with a bit of a hack: the filter
basically matches the window title against "^i3bar for output.*".
There's currently no other (and cleaner) way to do this.