Remove -fn, -p, -nb, -nf, -sb, -sf options in favor of a custom command
This removes a lot of code and complexity and is way more flexible.
I have multiple versions of dmenu with different patches installed,
that have different names. Not hardcoding the command is more flexible
in these situation, or if one wants to filter quickswitchs' output.

I would suggest to merge this together with a major version change, as
the invocation changed, although migrating is very straightforward.